requirements.txt
numpy
torch
torchvision
scikit-image
Pillow
